pyproject.toml
[tool.poetry]
name = "study-db"
version = "0.1.0"
description = ""
authors = ["vsouth <neppchela@gmail.com>"]
readme = "README.md"
packages = [{include = "study_db"}]

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.19"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
